+++ ChromeExtension/test/.babelrc
@@ -1,0 +1,13 @@
+
+{
+    "presets": [
+      [
+        "@babel/preset-env",
+        {
+          "targets": {
+            "node": true
+          }
+        }
+      ]
+    ]
+  }
